ExpenceTracker/ExpenseTracker/requirements.txt
Django==1.10
django-cors-headers==2.0.2
djangorestframework==3.6.3
